tech: mutualisa° du fetch des structures
    Modification de quelques modèles organization pour retourner un Organization complet plutôt qu'un OrganizationRaw. L'objectif est de disposer d'une fonction modèle capable de retourner un objet organization complet sur la base d'un id.
