+++ packages/api/server/models/organizationModel/findByIds.ts
@@ -1,21 +1,5 @@
-import { sequelize } from '#db/sequelize';
-import { QueryTypes } from 'sequelize';
+import { type Transaction } from 'sequelize';
+import find from './_common/find';
+import { Organization } from '#root/types/resources/Organization.d';
 
-export default ids => sequelize.query(
-    `SELECT
-        organizations.organization_id AS id,
-        organizations.name,
-        organizations.fk_type,
-        organizations.being_funded,
-        organizations.being_funded_at,
-        organization_types.fk_category
-    FROM organizations
-    LEFT JOIN organization_types ON organizations.fk_type = organization_types.organization_type_id
-    WHERE organizations.organization_id IN (:ids)`,
-    {
-        type: QueryTypes.SELECT,
-        replacements: {
-            ids,
-        },
-    },
-);
+export default (ids: number[], transaction?: Transaction): Promise<Organization[]> => find({ ids }, transaction);
